Create the initial home-league assignment algorithm
The first major piece of functionality that we need for team assignment is
the ability to find the best set of "home" league assignments for every
team in Washington. This commit has two different approaches to doing that

1. Use a heuristic algorithm to assign teams to their closest events in
   a randomized order, running that many times to find the best result
   (determined by the total sum of distance traveled).
2. Group teams by zip code, and assign the groups of teams to events starting
   with the densest zip codes (to avoid locking them out).

After doing some testing, approach 2 provides more promising results, while
also running faster, and allows for easier tweaking if teams are slotted to
a non-optimal league.

In either approach, the code is designed to support pre-assigning teams to events
to cover a variety of use cases (hosts, teams where geography means their cloesest
event as the crow flies is wrong, or when an override is needed).

The output is a pair of files, one CSV with league name and team number, and a second
CSV that has coordinates that can be imported into a Google Map for easier visualization.