Add league summary print out
One of the things we want to look at whenever we're done moving teams
around is what the distribution looks like across all of the leagues
(beacuse we don't want a league of 5 if we can help it). So now we have
a function to do a really basic printout of each league and it's
capacity status.