Improve League handling of teams
While it makes sense to tie a new team in a league to a distance when
looking at home leagues (for league analysis), it makes less sense when
running through preferences. So that field is now optional on the model.

Separately we also need to be able to remove teams from a league when
doing preferencing, so there's a new method for that.