Add initial preference-handling script
This adds the `process_event_preferences.py` script, which is designed
to overlay a team's event preferences over the originally-assigned home
leagues in order to create the final event assignments for the year.

The key logic is that we take in the preferences, remove those teams
from their home leagues, and then work through the preferences in a
random order (seeded to the year so it's deterministic). We go through
each team's preference list in order, and if all of those events are
full we place them back into their home league (regardless of capacity).
If this results in a league becoming too large, we can achieve the same
effect without going over size by removing the team that couldn't be
placed from the preference submissions (so that they are never removed
from their home event, and someone isn't placed into "their" spot).